requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Medical Imaging
nibabel>=5.0.0
SimpleITK>=2.2.0
pylidc>=0.2.3
pydicom>=2.4.0

# nnU-Net
nnunetv2>=2.3.0

# Data & Utilities
pandas>=2.0.0
scikit-learn>=1.2.0
scikit-image>=0.20.0
tqdm>=4.65.0
pyyaml>=6.0
h5py>=3.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Metrics
medpy>=0.4.0
surface-distance>=0.1

# Experiment Management
tensorboard>=2.12.0
wandb>=0.15.0

# Testing
pytest>=7.3.0
